% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHierarchicalDataObject.R
\name{addHierarchicalDataObject}
\alias{addHierarchicalDataObject}
\title{Recursively adding an entire folder to a data package}
\usage{
addHierarchicalDataObject(folder_path, pattern, pkg, metadataId)
}
\arguments{
\item{folder_path}{(character) where the files are}

\item{pattern}{(character) the extension to remove}

\item{pkg}{data package}

\item{metadataId}{the metadata idenfitier}
}
\value{

}
\description{
Recursively adding an entire folder to a data package
}
