% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_att.R
\name{annotate}
\alias{annotate}
\title{Constructs an EML annotation for attributes}
\usage{
annotate(label, valueURI)
}
\arguments{
\item{label}{character}

\item{valueURI}{URI}
}
\value{
EML annotation
}
\description{
Constructs an EML annotation for attributes
}
