% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_att.R
\name{annotate_att}
\alias{annotate_att}
\title{Add attribute annotation to eml document}
\usage{
annotate_att(d, index = NULL, attribute, identifier, label, valueURI)
}
\arguments{
\item{d}{eml document}

\item{index}{integer a number}

\item{attribute}{character - attributeName}

\item{identifier}{character - a unique identifier}

\item{label}{character - the value semantic annotation label}

\item{valueURI}{character- value URI}
}
\value{

}
\description{
Add attribute annotation to eml document
}
