% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalizes Habitat}
\usage{
capitalize(ds)
}
\arguments{
\item{ds}{dataframe}
}
\value{
dataframe
}
\description{
Capitalizes the first word in the herbarium spreadsheet column habitat
}
