% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHierarchicalDataObject.R
\name{createDataObject}
\alias{createDataObject}
\title{Create new dataObjects based on directory}
\usage{
createDataObject(file_name, folder_path, pattern, pkg)
}
\arguments{
\item{file_name}{(character)}
}
\value{

}
\description{
Create new dataObjects based on directory
}
