% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAIr-package.R
\docType{package}
\name{LAIr-package}
\alias{LAIr}
\alias{LAIr-package}
\title{LAIr: Converting NDVI to LAI of Field, Proximal and Satellite Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Convert Leaf Area Index (LAI) from the Normalized Difference Vegetation Index (NDVI) using available equations from literature. Detailed description of conversion equations in Bajocco et al. 2022 \doi{10.3390/rs14153554}.
}
\author{
\strong{Maintainer}: Francesco Chianucci \email{fchianucci@gmail.com} (\href{https://orcid.org/0000-0002-5688-2060}{ORCID})

Authors:
\itemize{
  \item Francesco Savian \email{francesco.savian@vivairauscedo.com}
  \item Carlotta Ferrara \email{carlottaferrara@gmail.com} (\href{https://orcid.org/0000-0001-5249-1196}{ORCID})
  \item Fabrizio Ginaldi
  \item Nicola Puletti \email{nicola.puletti@gmail.com} (\href{https://orcid.org/0000-0002-2142-959X}{ORCID})
  \item Simone Bregaglio \email{simone.bregaglio@crea.gov.it} (\href{https://orcid.org/0000-0001-8381-2626}{ORCID})
  \item Sofia Bajocco \email{sofiabajocco@gmail.com} (\href{https://orcid.org/0000-0003-2301-9188}{ORCID})
}

}
\keyword{internal}
