% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tree.R
\name{print_tree}
\alias{print_tree}
\title{Print tree with number of files}
\usage{
print_tree(path, level = 1)
}
\arguments{
\item{path}{the file path}

\item{level}{number of folder levels recursively to go down}
}
\value{

}
\description{
modified from the function [dir_tree](https://fs.r-lib.org/reference/dir_tree.html) from the [fs package](https://fs.r-lib.org/index.html)
}
